Refactor product form layout management and streamline image field handling
